requirements.txt
breeze-connect>=1.0.34
pandas>=1.3.0
numpy>=1.20.0
pyotp>=2.6.0
python-dotenv>=0.19.0
h5py>=3.0.0
sqlalchemy>=1.4.0
matplotlib>=3.4.0
seaborn>=0.11.0
fastapi>=0.68.0
uvicorn>=0.15.0
click>=8.0.0
textual>=0.11.0
pytest>=6.0.0